requirements.txt
opencv-python
imagesize
numpy
Pillow
tqdm
